+++ src/decorators/params/SimpleCommandParams.ts
@@ -1,11 +1,11 @@
-import { Snowflake } from "discord.js";
+import { IGuild } from "../..";
 
 export interface SimpleCommandParams {
   argSplitter?: string | RegExp;
   description?: string;
   directMessage?: boolean;
   defaultPermission?: boolean;
-  guilds?: Snowflake[];
+  guilds?: IGuild[];
   botIds?: string[];
   aliases?: string[];
 }