+++ src/index.ts
@@ -4,3 +4,4 @@ export * from "./decorators";
 export * from "./types";
 export * from "./Client";
 export * from "./classes";
+export * from "./util";